move models to separate file; cleanup
